Alterado forma de busca das partidas em andamento com filtro de liga e time. Agora é necessário informar o ID da liga ou do time e não mais o nome. Enfrentei problemas com ligas com nome repetido que acabaram gerando alguns problemas no programa, portanto resolvi deixar a busca pelo ID para ser mais acertiva
